refactor (styles): cleanup and overflow fix
