+++ src/popup/screens/Details/DetailsHeader.css.ts
@@ -3,11 +3,6 @@ import { style } from "@vanilla-extract/css";
 export const detailsHeaderTitle = style({
   justifyContent: 'center',
   display: 'flex',
-  alignItems: 'center'
+  alignItems: 'center',
+  gap: 4,
 })
-
-// TODO: fix
-// export const detailsHeaderFavicon = style({
-//   margin: "4px",
-//   marginRight: "8px"
-// })
