+++ src/popup/screens/Overview/OverviewHeader.css.ts
@@ -2,7 +2,9 @@ import { style } from "@vanilla-extract/css";
 import { headerContentBase } from "../../components/DetailHeader.css";
 import { headerTitleBase } from "../Latest/LatestHeader.css";
 
-export const overviewHeaderContainer = style({})
+export const overviewHeaderContainer = style({
+  marginTop: 4,
+})
 
 export const overviewHeaderContent = style([headerContentBase, {
   gridTemplateColumns: "auto 40px"
