+++ src/popup/screens/Overview/OverviewScreen.css.ts
@@ -2,4 +2,5 @@ import { style } from "@vanilla-extract/css";
 
 export const screenList = style({
   marginTop: 4,
+  overflowY: "auto",
 })
