+++ src/styles/chrome/_vars.scss
@@ -26,16 +26,3 @@ $li-count-hover-color: var(--color-font);
 
 // Buttons
 $bookmark-close-size: 60%;
-
-// @mixin item-buttons {
-//   @include item-button(".remove", $removeIcon, 70%);
-//   @include item-button(".bookmark", $bookmarkIcon, 60%);
-//   @include item-button(".bookmark-all", $bookmarkAllIcon, 60%);
-//   @include item-button(".get-in", $chGetInIcon, 40%, inherit, inherit);
-// }
-
-// @mixin header-buttons {
-//   @include header-button(".close-all", $closeAllIcon, 50%);
-//   @include header-button(".search-btn", $searchBtnIcon, 50%);
-//   @include header-button(".back", $goBackIcon, 45%, var(--color-back-button));
-// }
