--- src/styles/common/_animations.scss
@@ -1,34 +1,0 @@
-/* ANIMATIONS */
-.slide-in {
-    transform: translate(0) !important;
-
-    transition: transform 150ms cubic-bezier(.07, .95, 0, 1) !important;
-}
-
-.slide-out {
-    transform: translate(-100%);
-
-    transition: transform 150ms cubic-bezier(.07, .95, 0, 1);
-    position: relative;
-}
-
-#overview.slide-out {
-    overflow-y: hidden !important;
-}
-
-.slide-out-reverse {
-    transform: translate(100%);
-
-    transition: transform 150ms cubic-bezier(.07, .95, 0, 1);
-}
-
-.slide-in-reverse {
-    transform: translate(0);
-
-    transition: transform 150ms cubic-bezier(.07, .95, 0, 1);
-}
-
-#overview.slide-in-reverse {
-    overflow-y: auto !important;
-}
-
