--- src/styles/common/_icons.scss
@@ -1,15 +1,0 @@
-// Buttons and icons
-$bookmarkIcon: url(../../icons/btn-bookmark-star.svg);
-$bookmarkCloseIcon: url(../../icons/ico-bookmark-close.svg);
-$bookmarkAllIcon: url(../../icons/ico-bookmark-close.svg);
-$bookmarkedIcon: url(../../icons/btn-bookmark-star-full.svg);
-
-$goBackIcon: url(../../icons/btn-big-arrow.svg);
-$closeAllIcon: url(../../icons/btn-big-remove.svg);
-
-$getInIcon: url(../../icons/btn-bold-arrow.svg);
-$removeIcon: url(../../icons/btn-remove.svg);
-$searchBtnIcon: url(../../icons/btn-search-24.svg);
-
-// Chrome-only
-$chGetInIcon: url(../../icons/btn-arrow-filled-ch.svg);
