--- src/styles/common/_mixins.scss
@@ -1,69 +1,0 @@
-// .close-all, #search-btn (.back later)
-// @mixin header-button ($btn-name, $img, $img-size, $bck-color: inherit, $hover-color: var(--button-hover-color), $active-color: var(--button-active-color)) {
-//   #{$btn-name}.header-button {
-//     width: var(--large-button-dimension);
-//     height: var(--large-button-dimension);
-//     border-radius: var(--border-radius);
-
-//     border: 0; // CHECKED
-//     padding: 0;
-
-//     background-color: $bck-color;
-
-//     display: flex;
-//     justify-content: center;
-//     align-items: center;
-
-//     color: rgba(12, 12, 13);
-
-//     @media (prefers-color-scheme: dark) {
-//       color: rgba(249, 249, 250, 0.8); // TODO fix opacity!
-//     }
-
-//     @if $btn-name==".back" {
-//       transform: rotate(180deg);
-//     }
-
-//     &:hover {
-//       background-color: $hover-color;
-//     }
-
-//     &:active {
-//       background-color: $active-color;
-//     }
-//   }
-// }
-
-// @mixin item-button($btn-name, $img, $img-size, $hover-color: var(--button-hover-color), $active-color: var(--button-active-color)) {
-//   #{$btn-name} {
-//     width: var(--small-button-dimension);
-//     height: var(--small-button-dimension);
-//     border-radius: var(--border-radius);
-//     flex-shrink: 0;
-
-//     border: 0; // CHECKED
-
-//     background-image: $img;
-//     background-color: inherit;
-
-//     color: rgba(12, 12, 13);
-
-//     @media (prefers-color-scheme: dark) {
-//       color: rgba(249, 249, 250, 0.8); // TODO fix opacity!
-//     }
-
-//     &:hover {
-//       background-color: $hover-color;
-//     }
-
-//     &:active {
-//       background-color: $active-color;
-//     }
-//   }
-// }
-
-@mixin truncate-text {
-  overflow-x: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
