+++ src/styles/common/style_common.scss
@@ -1,7 +1,3 @@
-@import "mixins";
-@import "animations";
-@import "search";
-
 // HEADER
 
 // Header buttons
